share plotWidget fixture
